+++ core/src/ZXBigInteger.cpp
@@ -457,7 +457,7 @@ BigInteger::Multiply(const BigInteger &a, const BigInteger &b, BigInteger& c)
 static int GetBitLen(uint32_t x)
 {
 	return x == 0 ? 0 : BitHacks::HighestBitSet(x) + 1;
-	size_t len = 0;
+	int len = 0;
 	while (x > 0) {
 		x >>= 1;
 		len++;
