+++ core/src/oned/ODCode93Reader.cpp
@@ -207,7 +207,7 @@ DecodeExtended(const std::string& encoded, std::string& decoded)
 static int IndexOf(const char* str, char c)
 {
 	auto s = strchr(str, c);
-	return s != nullptr ? (s - str) : -1;
+	return s != nullptr ? static_cast<int>(s - str) : -1;
 }
 
 bool
